Add file transfer to shell and attach views
