+++ b4n/core/managers/views.rs
@@ -554,11 +554,11 @@ impl ViewsManager {
 
     /// Opens shell / attach to the main process of the specified container.
     pub fn open_shell(&mut self, resource: ResourceRef, is_attach: bool) {
-        if let Some(client) = self.worker.borrow().kubernetes_client() {
+        if let Some(client) = self.worker.borrow().kubernetes_client().map(|c| c.get_client()) {
             self.footer().hide_hint();
             let view = ShellView::new(
-                self.worker.borrow().runtime_handle().clone(),
                 Rc::clone(&self.app_data),
+                Rc::clone(&self.worker),
                 client,
                 resource.into(),
                 is_attach,
