+++ b4n/ui/views/common.rs
@@ -1,8 +1,10 @@
 use b4n_common::NotificationSink;
+use b4n_kube::{ContainerType, ResourceTag};
 use b4n_tasks::commands::CommandResult;
 use b4n_tui::{ResponseEvent, TuiEvent};
 use ratatui::Frame;
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
+use std::path::PathBuf;
 use std::rc::Rc;
 use std::time::{Duration, Instant};
 use tui_term::vt100::Screen;
@@ -100,6 +102,48 @@ pub fn get_layout_with_header(area: Rect) -> Rc<[Rect]> {
         .direction(Direction::Vertical)
         .constraints(vec![Constraint::Length(1), Constraint::Fill(1)])
         .split(area)
+}
+
+/// Gets names for all containers in the list of resource tags.
+pub fn get_all_containers_from_resource_tags(tags: &[ResourceTag]) -> Vec<String> {
+    tags.iter()
+        .filter_map(|t| match t {
+            ResourceTag::Container(name, _, _) => Some(name.to_ascii_lowercase()),
+            _ => None,
+        })
+        .collect()
+}
+
+/// Gets names for all non-ephemeral containers in the list of resource tags.
+pub fn get_target_containers_from_resource_tags(tags: &[ResourceTag]) -> Vec<String> {
+    tags.iter()
+        .filter_map(|t| match t {
+            ResourceTag::Container(name, kind, _) if *kind != ContainerType::Ephemeral => Some(name.to_ascii_lowercase()),
+            _ => None,
+        })
+        .collect()
+}
+
+/// Tries to guess selected path for the file picker basing on the `is_download` flag.
+pub fn get_path_for_file_picker(path: &str, is_download: bool) -> PathBuf {
+    fn default_path() -> PathBuf {
+        std::env::current_dir().unwrap_or(PathBuf::from("."))
+    }
+
+    if path.trim().is_empty() {
+        default_path()
+    } else {
+        let path = PathBuf::from(path);
+        if is_download {
+            path
+        } else if let Some(path) = path.parent()
+            && !path.as_os_str().is_empty()
+        {
+            PathBuf::from(path)
+        } else {
+            default_path()
+        }
+    }
 }
 
 /// Allowed time between two key presses that will still activate the escape sequence mode.
