+++ b4n/ui/views/resources/mod.rs
@@ -4,5 +4,4 @@ pub use self::view::*;
 mod dialogs;
 mod menus;
 mod table;
-mod utils;
 mod view;
